requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
